+++ backend-node/data/index.js
@@ -1,11 +1,10 @@
 const crypto = require('crypto');
+const { Pool } = require('pg');
 
-const { Client } = require('pg');
 const connectionString = process.env.postgresDatabaseUrl;
-const client = new Client({
+const pool = new Pool({
   connectionString
 });
-
 const passwordSalt = process.env.salt;
 
 const createUser = async function (username, password, name) {
@@ -13,13 +12,9 @@ const createUser = async function (username, password, name) {
   let user;
 
   try {
-
-    await client.connect();
-
     user = await findUser(username);
 
     if (user) {
-      await client.end();
       user.status = 'existed';
       return user;
     }
@@ -29,8 +24,7 @@ const createUser = async function (username, password, name) {
       .update(password)
       .digest('hex');
 
-    const res = await client.query('INSERT INTO users(name, username, password) VALUES($1, $2, $3) RETURNING name, username', [name, username, passwordHash]);
-    await client.end();
+    const res = await pool.query('INSERT INTO users(name, username, password) VALUES($1, $2, $3) RETURNING name, username', [name, username, passwordHash]);
 
     if (res.rowCount === 1) {
       user = res.rows[0];
@@ -55,9 +49,7 @@ const identifyUser = async function (username, password) {
       .update(password)
       .digest('hex');
 
-    await client.connect();
-    const res = await client.query('SELECT name, username from users where username=$1::text and password=$2::text', [username, passwordHash]);
-    await client.end();
+    const res = await pool.query('SELECT name, username from users where username=$1::text and password=$2::text', [username, passwordHash]);
 
     if (res.rowCount === 1) {
       user = res.rows[0];
@@ -75,10 +67,9 @@ const findUser = async function (username) {
   let user;
 
   try {
-
     username = username.toLowerCase();
 
-    const res = await client.query('SELECT name, username from users where username=$1::text', [username]);
+    const res = await pool.query('SELECT name, username from users where username=$1::text', [username]);
 
     if (res.rowCount === 1) {
       user = res.rows[0];
